Generalize pattern matching implementation
These changes are meant to lay the foundations for future developments. No new behavior has been introduced.

The grammar and interpreter have been modified in order to clearly separate the concepts of pattern matching & query evaluation.

The most noticeable changes are located in the grammar. The modifications on the interpreter are mostly relocating some existing code.