To_Unbounded_Text ambiguities
There is lots of ambiguities due to the To_Unbounded_Text overload
especially when using literals, that can both be converted to
String and to Text_Type. In order to solve those I used two solutions:
1. Text_Type qualification
2. To_Unbounded_Text (T: Text_Type) renaming: this works because only
one version of the function is renamed hence the result is not overloaded.

Both are quite ugly and really a bit disappointing.