Fix the project relative file fetching
This function was wrongly implemented because it was concatenating the
CLI provided '-P' option with the required file-name. Thus, when using a
project located in the 'GPR_PROJECT_PATH', the file fetching was invalid.

This fix is required for later commit in the same MR.

(cherry picked from commit 2a77485f8d5ca30ba7e3e24d08f377a3e98e8d52)