Genericize LKQL: workaround call to non-static method (another occurrence)
This is a workaround that should be fixed to make LKQL generic.

The issue is that we cannot access to static method through Java
interfaces nor abstract classes (at least to my knowledge). Here, the
static method is required to have access to an object of the same
class, representing a "null" object. Why can't we just call a
constructor instead?