Fix build with secure control plane
Since GraalVM version 24 is used, the `lkql` build breaks when the
_Secure Control Plane_ is enabled on `rocky8`.

The error message on failure is:
```
_[engine] WARNING: The Truffle API JAR is missing the 'truffleattach' resource, likely due to issues when Truffle was repackaged into a fat JAR.
_As a result, the optimized Truffle runtime is unavailable, and Truffle cannot provide native access to languages and tools.
_To customize the behavior of this warning, use the 'polyglotimpl.AttachLibraryFailureAction' system property.
```

The root cause of the issue is described by the graalvm issue
[8222](https://github.com/oracle/graal/issues/8222), where the
`polyglot` package tries to create a `.cache` directory in the
`user.home` directory.

When using the _Secure Control Plane_, only the buildspace of the spec
is a writable directory.

Setting the `polyglot.engine.userResourceCache` variable to a writable
path fixed the problem, and it is now possible to build with
_Secure Control Plane_ enabled.

Closes eng/libadalang/langkit-query-language#549+