Fallback on a 'dumb' terminal when failing to create a system one
We use the 'JLine' Java library to implement the LKQL REPL, and the latter use an
abstraction over terminals. In order to avoid warning message on not fully supported
terminals, we have to tell the library to fallback on a 'dumb' one.
See https://jline.org/docs/terminal for more info about JLine.